09_palindromes: Amend array mutation bug
The line that tried to reverse the array was reversing it every iteration, and it was mutating the same array too.
It is better to have a separate variable hold the reversed array, and to use .slice() to make sure it is a new array
and not a mutation of the same array reference!